# Conflicts:
#	AuctionKoi.sql
#	pom.xml
#	src/main/java/swp/auctionkoi/configuration/ApplicationInitConfig.java
#	src/main/java/swp/auctionkoi/configuration/SecurityConfig.java
#	src/main/java/swp/auctionkoi/controller/AuctionController.java
#	src/main/java/swp/auctionkoi/controller/AuthenticationController.java
#	src/main/java/swp/auctionkoi/controller/StaffController.java
#	src/main/java/swp/auctionkoi/controller/UserController.java
#	src/main/java/swp/auctionkoi/exception/ErrorCode.java
#	src/main/java/swp/auctionkoi/models/Auction.java
#	src/main/java/swp/auctionkoi/models/Bid.java
#	src/main/java/swp/auctionkoi/models/Delivery.java
#	src/main/java/swp/auctionkoi/models/Transaction.java
#	src/main/java/swp/auctionkoi/models/enums/DeliveryStatus.java
#	src/main/java/swp/auctionkoi/repository/AuctionParticipantsRepository.java
#	src/main/java/swp/auctionkoi/repository/AuctionRequestRepository.java
#	src/main/java/swp/auctionkoi/repository/BidRepository.java
#	src/main/java/swp/auctionkoi/repository/TransactionRepository.java
#	src/main/java/swp/auctionkoi/repository/WalletRepository.java
#	src/main/java/swp/auctionkoi/service/auction/AuctionService.java
#	src/main/java/swp/auctionkoi/service/auction/impl/AuctionServiceImpl.java
#	src/main/java/swp/auctionkoi/service/auctionrequest/impl/AuctionRequestServiceImpl.java
#	src/main/java/swp/auctionkoi/service/payment/impl/PaymentServiceImpl.java
#	src/main/java/swp/auctionkoi/service/user/ManagerService.java
#	src/main/java/swp/auctionkoi/service/user/impl/ManagerServiceImpl.java
#	src/main/java/swp/auctionkoi/service/wallet/WalletService.java
#	src/main/java/swp/auctionkoi/service/wallet/impl/WalletServiceImpl.java

+++ src/main/java/swp/auctionkoi/models/AuctionRequest.java
@@ -35,13 +35,13 @@ public class AuctionRequest {
     swp.auctionkoi.models.Auction auction;
 
     @Column(name = "buy_out")
-    Double buyOut;
+    Float buyOut;
 
     @Column(name = "start_price")
-    Double startPrice;
+    Float startPrice;
 
     @Column(name = "increment_price")
-    Double incrementPrice;
+    Float incrementPrice;
 
     @Column(name = "method_type")
     Integer methodType;
